enabled absolute paths in project and reorganized a bit
