+++ client/src/store.js
@@ -1,5 +1,5 @@
 import { configureStore } from '@reduxjs/toolkit'
-import homeReducer from './screens/home/homeSlice'
+import homeReducer from 'screens/Home/homeSlice'
 
 export default configureStore({
   reducer: {
